+++ src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/core/AggregateExpression.kt
@@ -23,17 +23,3 @@ interface AggregateExpression<T>{
         }
     }
 }
-
-fun <T, R> AggregateExpression<T>.map(transform: (T) -> R): AggregateExpression<R> {
-    return AggregateExpression.of { context, path ->
-        this.compute(path, context).map { export -> export.map(transform) }
-    }
-}
-
-fun <T1, T2, R> combine(exp1: AggregateExpression<T1>, exp2: AggregateExpression<T2>, transform: (T1, T2) -> R): AggregateExpression<R> {
-    return AggregateExpression.of { context, path ->
-        combine(exp1.compute(path, context), exp2.compute(path, context)) { result1, result2 ->
-            ExportTree(transform(result1.root, result2.root))
-        }
-    }
-}
