+++ src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/core/Context.kt
@@ -1,5 +1,6 @@
 package io.github.filippovissani.kotlin_distributed_dfrp.core
 
+import io.github.filippovissani.kotlin_distributed_dfrp.core.impl.ContextImpl
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.runBlocking
 
@@ -18,27 +19,5 @@ interface Context {
         operator fun invoke(selfID: DeviceID, sensors: Map<SensorID, *> = emptyMap<SensorID, Any>()): Context {
             return ContextImpl(selfID, sensors)
         }
-    }
-}
-
-internal class ContextImpl(override val selfID: DeviceID, sensors: Map<SensorID, *>) : Context {
-    private val _neighborsStates = MutableStateFlow(emptyMap<DeviceID, Export<*>>())
-    private val _sensorsStates = sensors.map { (k, v) -> k to MutableStateFlow(v) }.toMap()
-    override val neighbors = _neighborsStates.asSharedFlow()
-
-    override fun receiveExport(neighborID: DeviceID, exported: Export<*>) {
-        runBlocking {
-            _neighborsStates.update { it.plus(neighborID to exported) }
-        }
-    }
-
-    override fun <T> updateLocalSensor(sensorID: SensorID, newValue: T) {
-        runBlocking {
-            _sensorsStates[sensorID]?.update { newValue }
-        }
-    }
-
-    override fun <T> sensor(sensorID: SensorID): Flow<T> {
-        return _sensorsStates[sensorID]?.asStateFlow() as Flow<T>
     }
 }
