+++ src/main/kotlin/io/github/filippovissani/kotlin_distributed_dfrp/core/impl/ContextImpl.kt
@@ -1,0 +1,30 @@
+package io.github.filippovissani.kotlin_distributed_dfrp.core.impl
+
+import io.github.filippovissani.kotlin_distributed_dfrp.core.Context
+import io.github.filippovissani.kotlin_distributed_dfrp.core.DeviceID
+import io.github.filippovissani.kotlin_distributed_dfrp.core.Export
+import io.github.filippovissani.kotlin_distributed_dfrp.core.SensorID
+import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.runBlocking
+
+internal class ContextImpl(override val selfID: DeviceID, sensors: Map<SensorID, *>) : Context {
+    private val _neighborsStates = MutableStateFlow(emptyMap<DeviceID, Export<*>>())
+    private val _sensorsStates = sensors.map { (k, v) -> k to MutableStateFlow(v) }.toMap()
+    override val neighbors = _neighborsStates.asSharedFlow()
+
+    override fun receiveExport(neighborID: DeviceID, exported: Export<*>) {
+        runBlocking {
+            _neighborsStates.update { it.plus(neighborID to exported) }
+        }
+    }
+
+    override fun <T> updateLocalSensor(sensorID: SensorID, newValue: T) {
+        runBlocking {
+            _sensorsStates[sensorID]?.update { newValue }
+        }
+    }
+
+    override fun <T> sensor(sensorID: SensorID): Flow<T> {
+        return _sensorsStates[sensorID]?.asStateFlow() as Flow<T>
+    }
+}
